refactor: improve country subdivisions options preparation with util function
